Move Payload impl into protocol package.
