+++ internal/protocol/common.go
@@ -7,7 +7,8 @@ import (
 	"github.com/deadblue/elevengo/lowlevel/client"
 )
 
-func (c *ClientImpl) do(req *http.Request) (resp *http.Response, err error) {
+// |send| sends an HTTP request, returns HTTP response or an error.
+func (c *ClientImpl) send(req *http.Request) (resp *http.Response, err error) {
 	req.Header.Set("Accept", "*/*")
 	// Always override user-agent
 	ua := c.ua
@@ -30,6 +31,23 @@ func (c *ClientImpl) do(req *http.Request) (resp *http.Response, err error) {
 	return
 }
 
+// |post| performs an HTTP POST request to specific URL with given payload.
+func (c *ClientImpl) post(url string, payload client.Payload) (body io.ReadCloser, err error) {
+	req, err := http.NewRequest(http.MethodPost, url, payload)
+	if err != nil {
+		return
+	}
+	req.Header.Set(headerContentType, payload.ContentType())
+	if size := payload.ContentLength(); size > 0 {
+		req.ContentLength = size
+	}
+	var resp *http.Response
+	if resp, err = c.send(req); err == nil {
+		body = resp.Body
+	}
+	return
+}
+
 func (c *ClientImpl) Get(url string, headers map[string]string) (body io.ReadCloser, err error) {
 	var req *http.Request = nil
 	if req, err = http.NewRequest(http.MethodGet, url, nil); err != nil {
@@ -41,23 +59,7 @@ func (c *ClientImpl) Get(url string, headers map[string]string) (body io.ReadClo
 		}
 	}
 	var resp *http.Response = nil
-	if resp, err = c.do(req); err == nil {
-		body = resp.Body
-	}
-	return
-}
-
-func (c *ClientImpl) Post(url string, payload client.Payload) (body io.ReadCloser, err error) {
-	req, err := http.NewRequest(http.MethodPost, url, payload)
-	if err != nil {
-		return
-	}
-	req.Header.Set(headerContentType, payload.ContentType())
-	if size := payload.ContentLength(); size > 0 {
-		req.ContentLength = size
-	}
-	var resp *http.Response
-	if resp, err = c.do(req); err == nil {
+	if resp, err = c.send(req); err == nil {
 		body = resp.Body
 	}
 	return
